cmake: make glut a soft dependency
There's parts of the tree that is useful even without GLUT, so let's
allow building those without requiring GLUT. But let's emit a warning so
users have a chance to understand why they get such a stripped-down
build.

This is especially useful on Windows, where GLUT is a bit of an exotic
dependency.

This gets us a bit closer to what we're doing with autotools.

Acked-by: Martin Roukala <martin.roukala@mupuf.org>