tessdemo.c: cast glu tess callbacks
The gluTessCallback interface is kinda horrible, because it takes
pointers without really knowing what the correct signature is. Sadly,
this leads to compiler warnings, so let's cast the functions to quiet
these.

Let's use a typedef to keep this short and to the point, so we don't
need to repeat the same function prototype over and over again.