texaaline.c: clean up reused calculations
We calculated both sets of deltas for corordinates 2/4 and 3/5, even
though they were the same. But we only used the 2/3 set twice, ignoring
the 3/4 sets.

Let's instead rename these variables to show that they are valid for
both positions. That way we get less code, but it still makes sense.