add a meson build system
We're currently maintaining two separate build-systems, one based on
autotools, and one based on cmake. These are quite similar, but some
details are missing from each of them, most notably:

- The autotools build-system can't build the src/wgl sources
- The cmake build-system doesn't install things properly

There's a bunch of other smaller differences, but they're not that
important.

In theory we could have massaged the cmake build system to work well
enough for all use-cases, but there's a big elephan in the room here:

Mesa's main repository has moved to meson. So if we keep doing cmake,
the Mesa development community needs to learn two different build
systems that fills the same needs. That's not really ideal.

So instead, let's add a Meson build system, and later on remove the
other ones.

I think this is all; this should be a union of the feature-sets of the
autotools and cmake build systems.

Acked-by: Jose Fonseca <jfonseca@vmware.com>
Acked-by: Eric Engestrom <eric@engestrom.ch>