remove openvg support
The OpenVG state tracker was removed from Mesa back in 2015, over 7
years ago to date. The last release of an OpenVG version was back in
2008. It's pretty obvious that OpenVG is more or less dead.

So let's get rid of the OpenVG demos, as it seems they're not really
buildable or runnable on modern systems, at least not without a lot of
work.

Since the OpenVG code is the only code that uses Freetype2, this means
we can also yank out the code for that.

Reviewed-by: Chia-I Wu <olvaffe@gmail.com>