tests: remove getprocaddress test
While this test does *kinda* work, there's a few issues with it:
- It uses Python 2 rather than Python 3
- It doesn't build out-of-tree, because of some hard-coded paths
- It doesn't currently link, due to a missing X11 dependency

That last point especially seems to indicate that nobody has been using
this since the switch to autotools, which is a long time ago.

While all of the above can be fixed, we have since gotten Piglit, which
already extensively tests glXGetProcAddress. Let's get rid of it.

This also removes our last dependency on the mesa source code, so we can
now remove the logic for that as well.