autotools: distribute meson build-files
Without this, we can't really build meson from the autotools
distribution. So let's add these files!

In the case of the tools-folder, we also need to add dds.c to the
distrubution, because we actually build it in the meson build. In the
past, we only built it in the cmake-build, which we never added to the
distribution.