wincopy.c: do not copy potentially uninitialized member
We can't copy msg.wParam all the way out here, where we don't know if
the struct has actually been initialized or not. Let's instead copy it
each time we process a message.

But then we also need to stop overwriting it each time we get the
message, otherwise we'll end up with GetMessage's return value as the
final return value. Before we overwrote that variable with the wParam of
the last message in the end.