meson: fallback to glut-dependency
We should really prefer loading GLUT from a dependency, and then fall
back to looking for the header and library file if that fails instead.

Otherwise, we'll miss the system-installed GLUT on some systems, like
Debian, due to missing the include file in the default include path.

So let's try a dependency first, and fall back to the system-include if
not found.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>