meson: deprecate with-glut option
This approach has issues, where we can't pass the dependency we get from
declare_dependency to cc.has_function().

This flag was added to support the MinGW CI-build in the first place, but
that's now done using a pkg-config wrapper-file. So let's just mark it
as deprecated in case someone is using it, and move forward.

I wish we could use the "deprecated" keyword from Meson 0.60, but we
don't require a new enough version yet. Oh well, this isn't too bad
either.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>