add a mailmap-file
This mailmap-file is based on the output of "git shorlog -se", and then
taken duplicate entries and trying to figure out what the "preferred" or
current email address is. Most of those were picked from Mesa's
mailmap-file, which is likely to contain the contributors preference
already. Some were based on best-effort guesswork.

This allows us to see up-to-date-ish email addresses for old
contributions, which can be useful.

Acked-by: Hoe Hao Cheng <haochengho12907@gmail.com>