vkgears: fix non-gnu compilation
We can't use powf() as a compile-time function, which is required for a
global const array initializer.

So let's move this into the draw_gears() function instead. This is what
we do for the other arrays here as well, and a sufficiently clever
compiler should be able to tell that it's not being modified anyway.

Fixes: a92d6f62 ("vulkan: add vkgears demo")
Closes: https://gitlab.freedesktop.org/mesa/demos/-/issues/30
Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>