meson: enable some warnings on gcc/clang
This list was taken from Mesa, assuming it's a reasonable list to
ensure reasonable compatibility across compilers.

One difference from the list in mesa, is that I've changed
-Wimplicit-fallthrough to -Wimplicit-fallthrough=3. That change
makes this only enable on GCC, because Clang doesn't support
fallthrough comments like GCC does. Without this, we'd need to define
some FALLTHROUGH macro or something, because fallthrough annotations
aren't portable enough. So let's just punt on this, and get
fallthrough warnings on GCC only for now.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>