meson: do not search for glu on macos
On macOS, GLU is part of OpenGL.Framework, so we don't need to search
explicitly for it. Even if we did, we wouldn't find it, because we
include it by a different path than on other platforms.

So let's just bypass this entire check on macOS. We'll be fine anyway.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>