meson: disable some warnings on macos
On macOS, there's a difference between GLhandleARB and GLuint, which
means that the prototypes between GL_ARB_vertex_shader and the core
version of these APIs aren't strictly speaking compatible.

This isn't a real problem in reality, because nobody ends up with more
than 2^32-1 handles. But it makes the compiler complain.

So let's silence these warnings on macOS. It's going to be fine.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>