meson: explicitly check for glx dependency
X11 and GLX are really separate dependencies, and we can't assume GLX
is available just because X11 is. This matters on macOS, where GLX
isn't available unless xquartz is installed.

So let's test for glx seperately, so we know if we can use it or not.

For XQuartz-builds, we let libGL also provide GLX.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>