completely retire xmesa/fx code
This is ancient code that hasn't been hooked up to the build system
since the conversion to autoconf, back in 2007.

The code is here to support Mesa on Glide, which isn't something that
hasn't been supported in Mesa since Mesa 7.2/3, released in 2009. We
had a dummy implementation in Mesa that did nothing and always reported
failure until 2017, but now even that is gone.

It's about time to let this code go, as it has no chance to even build
on new Mesa versions.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>