meson: fix glx-detection without glvnd
When build without GLVND, Mesa provides GLX as a part of libGL, just
like XQuartz does on MacOS. There's nothing really macOS specific about
this fallback, so let's drop the condition.

And just for good measure, let's add in dep_gl as a dependency for
the header-check here, in case some one decides to install libGL outside
of the default include directory.

Fixes: 34300551 ("meson: explicitly check for glx dependency")
Closes: https://gitlab.freedesktop.org/mesa/demos/-/issues/36
Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>