vkgears: properly pick color-format
Before this, we just assumed our preferred color-format was available.
Now we instead look for it, and fall back to the first format if not
found.

We can probably do better in the future, though. Maybe we should prefer
format with more bits instead of having a single preferred format?

Acked-by: Hoe Hao Cheng <haochengho12907@gmail.com>