meson: bump minimum version to 0.60
We have limited our Meson version to what Mesa uses. But Mesa has
actually bumped their minimum Meson version to 0.60 in the mean time,
meaning we should be able to safely do the same.

This allows us to use the deprecated-keyword for options, instead of
open-coding our own. So let's do that while we're at it!

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>