egl/gl: Add an example of dma-buf texture sharing
The `dmabufshare` executable forks to create two processes, server and
client. The server creates a texture and shares it with the client.
Sharing is done with specific extensions from EGL and GL that use
DMA-BUF. The server periodically updates the texture to show that both
the server and the client reference the same texture.

Reviewed-by: Hoe Hao Cheng <haochengho12907@gmail.com>