vkgears: fix fences, use single precision, fix angular speed
Fix a bug in which vkAcquireNextImageKHR is called when back_buffer_semaphore
is set or is being waited on.

Use single precision floats when possible to avoid awkward double/float
mixing throughout the code.

Fix a bug in which the angular velocity of the gears increased during
resizing the window.

Signed-off-by: Emimendoza <emendoz@clemson.edu>